input date pick time
git-svn-id: http://192.168.0.243:4435/svn/bee/trunk@621 aaee961a-53e7-2249-bd01-b2a8f0c4ce08

+++ src/com/butent/bee/client/calendar/resourceview/ResourceViewHeader.java
@@ -61,7 +61,7 @@ public class ResourceViewHeader extends Composite {
       return;
     }
 
-    double width = 100d / attendees.size();
+    int width = 100 / attendees.size();
     for (int i = 0; i < attendees.size(); i++) {
       Label label = new Label(CalendarKeeper.getAttendeeName(attendees.get(i)));
       label.setStylePrimaryName("day-cell");
